cci/1: Simplify expression parser goals
I originally intended to support larger-than-word r-values in cci/1 to
implement 64-bit integers and by-value structs. This is possible with a
single-pass compiler but it would be much more complicated and not worth
it. It would also have required the 64-bit math and floating point
library functions in libc/1 which is not practical because omC doesn't
have unsigned.

There will be a bit of complexity supporting float and long long in
cci/2 but not nearly as much as the complexity involved in implementing
them in cci/1.