cci/1: Change va_list to point to next argument
The original va_list implementation pointed to the previous argument.
This is incompatible with cci/2, where it must point to the next
argument to properly handle arguments of different sizes.

In order to be able to link files compiled with cci/2 and cci/1
together, they need the same va_list ABI. e.g. when libc/2 is compiled
with cci/1 and linked with a program compiled by cci/2, printf() crashes
without this fix. This commit makes the va_list ABI the same (for those
types supported by cci/1.)