cci/1: Fix frame size of unbraced for loop
The for loop was not correctly tracking the maximum frame size before
calling locals_pop(); only parse_block() was, which meant an unbraced
for loop did not always have space for the iterator.

Frame size tracking has been moved into locals. It is much simpler now;
callers of locals_add() and locals_pop() do not need to know anything
about the frame size.