cci/2: Rewrite initializer list again
Parsing of an initializer list is now a simple loop in which we navigate
around the tree without recursion. It only recurses on nested braces.
This makes it much easier to understand.

Several bugs are fixed, most importantly initializing a struct inside
an initializer list with a scalar, as well as issues with braced
scalars.