cci/2: Emit ltu/lts instead of cmpu/cmps for comparisons
The cmpu and cmps instructions will be removed at some point. This is
probably slightly less efficient for now since cmpu is primitive and ltu
is not, but this will be better when ltu becomes primitive, and it makes
implementing long long comparisons easier so we might as well do it now.