Replace xor, ror with shl, shru
This removes the primitive instructions xor and ror, replacing them with
shl and shru. This commit affects all of the low-level components
including all VMs and anything written in bytecode or primitive
assembly.

Rotate right was originally used since it made it possible to implement
both shifts with a single instruction. Unfortunately, as it turns out,
it takes *many* instructions (nearly a dozen) to implement left and
right shift using rotate. On the other hand, xor and ror can be
implemented in few instructions (three each) and they are used much more
rarely. Shifts are critical for performance so they are now primitive
instructions.