hex/onramp: Replace cmpu with sub for equality
Same as the previous commit. This makes the code a bit harder to read
since sub is not obviously a comparison but it will be worth it to
simplify the use and implementation of the comparison instruction. There
is also the potential to better document this later by improving the
comparison instructions in compound assembly.