cci/0: Rename static initialization functions
The code previously used _init() for both static initialization of a
translation unit and initialization of an instance of a type. This was
not a problem for the early stages that have no non-trivial objects but
it became a problem in libo/1 and cci/1 where it's unclear what _init()
does.

This is part of a change to convert all static initialization to use
_setup()/_teardown() instead. Initialization of an instance will
continue to use _init()/_destroy(), and allocation and initialization
uses _new()/_delete(). It's much more clear now what each function does.