vm: Upgrade fread syscall test
This is the start of a series of changes to support a new mechanism for
syscalls. Syscalls will be provided as a table of function pointers
instead of a separate instruction, and syscalls will quash all numbered
and scratch registers. This will allow the Onramp OS to implement
syscalls as (mostly) ordinary functions and will allow programs to proxy
syscalls to child processes.

The first step in this process is to preserve registers when making
syscalls. This change preserves registers in the fread test. It also
properly reads the whole file in a loop and compares the entire
contents.