vm/c-debugger: Quash registers on syscall
This makes the debugger reset registers to bogus values on all syscalls.
The last few tests that were not working under it are now fixed.

The entire bootstrap process now preserves registers when making
syscalls. This will make it possible to implement syscalls as ordinary
functions in the Onramp OS. It also makes it possible to proxy syscalls
by child processes in the libc.