vm/python: Simplify and optimize opcode handling
The old code of generating a mask was meant to simplify things but it
really didn't work; it just made the code hard to understand. It's much
simpler to just do everything inline. This change also provides a ~30%
speedup (still much too slow to be useful for bootstrapping but it's
nice for testing at least.)

The downside is the Python VM has less error checking, but again, it's
too slow to be useful anyway.

This also adds Python 2.7 support for good measure.