hex/onramp: Convert to VM version 2, add buffering
Syscalls are now made via the syscall table with the external call
convention instead of the old sys instruction. The ftrunc syscall is
optional so the output file is only truncated if the syscall is
supported.

I/O is now buffered. This will greatly improve bootstrap performance on
VMs where syscalls are expensive.

Files are now closed properly before exiting.