cci/2: Fix self-referential variable initializers
These worked in cci/1 but not cci/2. This fixes it.

Variable declaration parsing is still very complicated in cci/2 to
handle the mess of tentative definitions, extern and static
declarations, shadowing declarations and so on. It would be nice to
clean up the code but I'm not sure if it can be made much simpler.