libc/2: Implement float addition
Currently only addition of non-negative finite floats is implemented. NaN,
infinity, and negative inputs are not implemented yet, but infinity is
produced correctly on overflow, and subnormal floats are handled
correctly. I believe it produces correctly rounded floats in all cases,
but only round-to-even is implemented so far.

Float parsing and formatting is not implemented yet so this can't really
be used by compiled programs yet. Float support is a work in progress.