vm: Add table counts to pit, bump version
This adds table lengths for the process info table and the syscall
table. This will allow them to be expanded with optional fields and
optional syscalls later without breaking compatibility with existing
VMs.

This is hopefully the last breaking VM change for a while since the spec
can now be expanded in a non-breaking way.

The alloc and free syscalls are reserved now even though nothing uses
them and no VMs implement them yet. This is done now so that once we
upgrade our bootstrapping programs to use them, programs will not need
to check the syscall table size before checking whether these syscalls
are implemented.

All of the fields in the process info table are now documented in their
own sections in the Onramp VM spec.