vm, libc: Separate program image from heap
The VM spec previously required the program image and free memory to be
contiguous. There was no point in having this restriction; it only made
program spawning more difficult to implement.

The "program break" field has been renamed to "heap start". It now
represents the start of free memory, and it is unrelated to the location
of the program. This does not break backwards compatibility (assuming no
programs actually used "program break" to find the end of their image.)

Programs can now spawn subprograms by allocating and loading the program
anywhere, then providing a separate region as free memory. This is much
simpler to implement, and spawn is faster and wastes less memory because
it does not need to copy the program image into contiguous subprogram
memory.