test/ld: Remove generated output for some tests
Tests no longer require matching to a generated output file. Instead a
.fail file is used to mark failing tests, the same as all the other
tools.

This fixes the problem of failing ld tests every time the libc changes
(see the parent commit.)

An eventual goal is to transition away from having any generated test
output in the repository. This is more difficult for the linker because
it produces binary output. Many of the test cases will need to be
rewritten to be runnable programs.