libc/2: Support NaN, inf, sign in float addition
Float addition now correctly handles NaN and infinity inputs. It also
handles sign correctly, except that subtraction is not implemented yet,
so adding floats of opposite signs does not work.

This also correctly implements quiet and signaling NaNs with the same
ABI as most platforms: the high bit of the significand is the quiet bit.
If a signaling NaN is passed to float addition or subtraction,
raise(SIGFPE) is called. If the signal returns, the NaN is quieted and
returned.