libc/3: Implement signal(), sigaction(), raise()
These were previously stubbed out. This is a somewhat more complete
implementation. Most flags are not implemented, but signal() works
correctly, and sigaction() works correctly and respects SA_SIGINFO and
SA_RESETHAND. Signal masking is not implemented yet.

Most signal types are still not raised by anything, but SIGFPE is raised
by signaling NaNs and can be caught, and resuming from a SIGFPE handler
works correctly (the NaN is quieted and returned.) A manual raise() of
other signals should work with a custom handler.