cg/0: Add first stage code generator
This is the first stage code generator. It is very much incomplete: it
doesn't have any optimizations yet but it correctly parses and outputs
assembly.

The build scripts have been updated so that it runs on itself and cpp/1.
It's not used to build any other tools yet; cc still needs to be updated
to run a code generator.