cci/0: Simplify function prologue
The entry point of each function is no longer in a separate _F_ symbol.
Instead the _F_ symbol contains only the frame size and the unprefixed
symbol is the entry point.

The old way was meant to be less instructions but that hasn't really
worked out and the result was ugly. This change makes functions easier
to understand and makes them easier to optimize with the code generator
(cg/0). It also eliminates all of the _F_ prefixes in debugger stack
traces.