cg/0: Improve push/pop optimization
The previous code did not inhibit push/pop optimizations around code
that used the stack, for example "ldw r0 rsp 0". cci/0 never emits code
like this but cci/1 does. cci/1 also uses r9 for its own purposes so we
can't use it in cg/0.

This commit adds some code to ensure r9 is not used for optimizations
and rsp is never touched inside of a push/pop optimization.