cci: Move test case runner into separate script
The cci test case runner (test/cci/test-case.sh) is now much more
powerful. It can parse //% metadata lines instead of using separate
files. It can run in parallel. It also places its build output in
predictable non-temporary locations so it can easily be inspected on
errors.

The overall batch run script (test/cci/run.sh) runs the new test case
script but is still mostly the same. The main difference is that it now
runs in the root of the repository (so there are no longer ../../.. all
over and commands can be run without changing directories.)

This work is done to support running cci test cases to test the stages
of the new cg codegen tools, and to eventually run all test cases in
parallel.