cg/0: Add support for running cci tests
The Makefile for cg/0 can now run it against all cci/0 tests in
conjuction with cci/0, as well as all cci/0 and cci/1 tests in
conjunction with cci/1.

The tests are not run by default in any of the test scripts yet because
there are some test failures with cci/0.