cg/0: Add constant folding
This performs the first part of constant folding optimizations. An
instruction that takes two constant inputs, or otherwise has a
degenerate case (such as multiplication by zero), is optimized to a
constant assignment.

This does not actually eliminate any instructions yet. Two other parts
still need to be implemented: constant propagation is only partially
complete and not activated yet; and a dead store elimination pass is
required to remove instructions that become redundant due to constant
folding.

The cg/0 code is a mess right now with lots of partially completed
stuff. It needs a lot of cleanup.