cg/0: Add propagation of adding instructions
For instructions that add arguments (ldw, stb, etc.), when one of the
arguments is zero and the other is a register, this propagates the
arguments of a previous add instruction into that register. This
optimizes many cases like array access where the compiler emits the
addition separately from the load/store.

This also fixes a bug with the horizon, a bug with the right shift
optimization, and a crash with mov from a function parameter. We also
now optimize libc/1.