cg/0: Document optimizations, mark complete
This does a bit of code cleanup but mostly just updates documentation.
Most of the optimizations were previously documented in code comments.
These have been moved to the cg/0 README.

cg/0 is now marked green on the main README; it is unlikely to see any
more additions. It is only used for bootstrapping, so additional
optimizations would probably take more time to run than they save. It
will probably undergo more cleanup if we add more features (like
structs) to cci/0, but otherwise it is done.