libc/0: Allocate FILE
A FILE is now a simple allocated structure with an allocated buffer. The
buffer is not used yet; code for buffering will be added next.

The code to close all open files on exit has been removed. It didn't
work anyway; it's not necessary during bootstrapping; and we're probably
going to relax the VM spec later to allow programs to exit without
closing files (and to optionally close the standard streams.)

The resulting code is much simpler because there is no longer a need to
maintain an array of all files.