cci: Detect dereference and subscript of void ptrs
Void pointer dereferences and subscripts are now detected correctly by
cci/1 and cci/2. cci/2 only warns against them to match GCC and Clang,
but using a dereferenced void pointer for assignments is now an error.
(Most other uses were already errors.)

cci/0 should probably include the same fix but for now I haven't
bothered. This could be done if and when we add the subscript operator
to it.

This also includes a fix to libc/3 strndup(). It worked by chance
before, but upcoming compiler improvements were causing it to compile to
nonsense due to an assignment to a subscripted void pointer.