sh: Implement `onrampvm` command
The shell can now run subprograms. Leaf build scripts now work except
for the `mkdir` commands which are not implemented yet, and `sh` is not
implemented yet either.

The shell still needs to proxy some file and directory syscalls in order
to make sure they are closed properly after the subprogram exits. There
will also be a lot of work to do to get alloc/free syscalls working
whenever we decide to add them.

The command_onrampvm() function is way too big right now; it needs to be
split up and part of it will be shared with command_sh().