docs/vm: Specify error codes for all syscalls
Two new error codes are added to clearly differentiate between a closed
stream and a non-blocking stream when no data is available for reading.
`ERROR_TRY_AGAIN` is returned to wait for data on a non-blocking stream
and `ERROR_END_OF_FILE` is returned at the end of a file or a closed
stream.

The error handling of all syscalls is now fully documented. In almost
all cases `ERROR_GENERIC` can be returned, but the use of additional
error codes is documented for more precise error handling.

See #24