libc/2: Add printf number flags, field width
This implements most of the flags, the field width, and most length
modifiers for integer formatting in printf().

This also simplifies the definition of intmax_t. It no longer has an
incompatible bootstrap definition; it is simply not defined during
bootstrapping.

printf() should be faster now because it uses only 32-bit math to format
numbers that are 32 bits or less. (64-bit formatting is mostly
implemented but is not entirely working yet.)