libc/2: Fix printf() %lld, add 64-bit modifiers
The %lld and %lli format specifiers are now fixed (some debug code was
accidentally left in.)

L is now a synonym for ll when used on integers. This is a GNU
extension.

The j modifier is now supported as another alias of ll (since intmax_t
is 64 bits on Onramp.)